refactor: Merge DisplayUnit component files
